+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # DuckDB Connector Changelog
 This changelog documents changes between release tags.
 
+## [0.1.1] - 2025-01-08
+* Add table comments to descriptions
+* Special case Timestamps and BigInt filtering
+
 ## [0.1.0] - 2024-08-22
 * Update Documentation for ndc-hub